change java version and improve classifieds search

+++ src/main/java/com/searchclassifieds/searchclassifieds/dao/ClassifiedRepository.java
@@ -1,7 +1,7 @@
 package com.searchclassifieds.searchclassifieds.dao;
 
-import com.searchclassifieds.searchclassifieds.dto.Classified;
+import com.searchclassifieds.searchclassifieds.model.Classified;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClassifiedRepository extends JpaRepository<Classified, Long> {
+public interface ClassifiedRepository extends JpaRepository<Classified, Integer> {
 }
